+++ src/main/java/pl/plajer/villagedefense/commands/arguments/admin/ClearEntitiesArgument.java
@@ -78,9 +78,14 @@ public class ClearEntitiesArgument {
               sender.sendMessage(registry.getPlugin().getChatManager().colorMessage("Kits.Cleaner.Nothing-To-Clean"));
               return;
             }
+            boolean eachThree = arena.getZombies().size() > 70;
+            int i = 0;
             for (Zombie zombie : arena.getZombies()) {
-              zombie.getWorld().spawnParticle(Particle.LAVA, zombie.getLocation(), 20);
+              if (eachThree && (i % 3) == 0) {
+                zombie.getWorld().spawnParticle(Particle.LAVA, zombie.getLocation(), 20);
+              }
               zombie.remove();
+              i++;
             }
             arena.getZombies().clear();
             arena.setOptionValue(ArenaOption.ZOMBIES_TO_SPAWN, 0);